requirements.txt
tensorflow
tensorflow_hub
numpy
pillow
gradio
